Add a custom roof
The final component for our procedural building set (for now) - a
roof!

This ended up being a lot more difficult than I initially conceived. I
wanted to have two different types of roof - a hip roof and a gabled
roof. I started with the hip roof because I (baselessly) assumed it
would be easier. It was not. In fact, trying to get the textures to
line up well was difficult enough that we may really be at the limit
of the procedural components.

So I ended up giving on the hip roof and just making the gable roof.
Even then, the design ended up changing a lot WHILE I was working on
it. The collision for the "roof-matter" (also called the fascia (I had
to learn a lot of roofing terms)) was particularly troublesome and
required a roundabout vertex stashing method.

Also I created an additional, more unusual export variable - a boolean
that controls whether we rebuild everything on an adjustment. This is
meant to cutdown on some lag that I'm experiencing in the editor.