Add a control to toggle generation in mesh
When everything was just singular procedural components, there was no
issues with the components updating every time a constraint was
changed or material was changed.

However, this behavior was starting to cause issues in the compound
components, like the Gap Wall or the Gap Set. In the case of both of
those, we were editing existing procedural components and adding new
ones, and then calling the build functions on all components. However,
since this code was running in the editor, the subcomponents were
rebuilding every time we updated a constraint. EVERY. SINGLE. VALUE.
UPDATE.

You can imagine how that would get out of hand - it wasn't too bad but
we were starting to see major issues with save times.

To counteract this, every subcomponent now has a boolean to toggle
mesh rebuild on a variable change. This should only be disabled if the
component in question is a subcomponent.