Add a roof to the house
Now that the roof is finally complete, I've added it to our working
house project.

The roof looks good, polygonally. It's got a good shape. But the
texture work looks... off. Unlike some of the other things in this
project I wouldn't call it straight-up reprehensible but it's
definitely not a highlight.