Finally texture the stupid stairs
After putting it off forever, I finally got around to texturing the
stairs for the house and the guardpost. The result?

Well, it looks good-to-okay (passable, really) but in order to
appropraitely texture each I had to create a whole slew of new
textures, each with a different name. I didn't come up with the best
naming scheme and even then I was inconsistent with how I executed it.

Having to create unique textures for each stair was a massive pain -
but everything was different for both stairs! The width, the length of
steps, the height of the steps. I wanted to highlight the edges so
the stairs had better definition, but since the edges were all over
the place, I had no choice but to make a bunch of separate textures. I
suppose it's not all that much trouble, but I like to keep my assets
nice and orderly!