Add a small hut with a thatched roof
As a counter-example to the wooden-shingled roof, I wanted to make a
smaller house with a thatched roof.

I've added it and called it the hut - while I appreciate the look of
the peaked roof and the square-ness of the building, the surface is
flat and boring. I'm really at my upper limit of what I can do with
textures and the like. The roofs either need to be Blender models
textured by someone who knows what they're doing, or just have
textures by someone who knows what they're doing. Not me, basically.

However, I don't think it'd be economical (or a good idea) to just
start commissioning models and textures without full scope of the work
to be done. So we'll leave it - FOR NOW!