Refactor scene & assets directories
As part of the asset reworks, we're gonna be adding a whole slew of
new scenes and assets, apportioned into different directories. Some
of the directories were growing fat (looking at you, formation!) so
I took it upon myself to re-organize everything.

While I'm more happy with the current form (buildings are now separate
from terrain, as are natural structures like rocks), the refactor was
actually REALLY destructive. Having all of these "tool" scripts
running in the background seems to have destroyed a bunch of scenes
and even just straight-up corrupted some stuff.

Fortunately I was able to fix everything, but I'm going to be much
more careful and considerate about when I do refactors of this
caliber.

Also, I deleted the building blocks, and the house made out of
building blocks. No need for that nonsense anymore.