Begin replacing the Surt-terrain with Fantasy24
For a while (it escapes me how long), we've been coloring our assets
using a pallette taken from Surt's Blowhard 2: Blowharder asset pack.
It was a sort of Gameboy Color-esque pallette that was bright and
poppy - I liked it a lot!

However, we bottomed out on what we (I) could do with it almost
immediately. A limited pallette is partially a blessing, partially a
curse: the limited color set looks good and helps me know what should
go where, but a pallette that's too limited stretches my flimsy art
skills to the limit.

I needed a pallette that had more tones - one that could express a
better range, and also trended towards warmer colors. The Fantasy24
pallette was the best one I could find. It has a good mix of colors,
all of them are warmer and (best of all) it comes with color usage
examples so that my tiny brain doesn't twist in twain. Neat!

The building textures were actually the first to use this pallette,
but this commit rolls out those changes to the basic terrain items.