Redo the trees in Fantasy24 pallette
I redid the crops and the grass; the next step was redoing the trees
in the Fantasy24 pallette.

This illustrated several issues to me. Firstly - we need to make trees
of our own. These current trees are old ones by SURT, and are too low
res for our current standard (32 pixels to a wold unit).

The other issue is the pallette. The problem with the original
pallette - which I believe was the NES pallette - was that it lacked
enough browns colors to appropriately show the wood-heavy setting
appropriately. The problem now is that the current pallette utterly
lacks the amount of greens required by NATURE, which is the game's
setting. While the yellowish highlights look really good on the trees,
they look pretty awful on the grass.