Replace Fantasy24 with Resurrect64 (modded)
As I outlined in the last commit, the Fantasy24 pallette had reached a
breaking point. While I liked the amount and variety of browns that
were provided, it basically had one green. One green does not a forest
make - maybe in the hands of a skilled artist, but definitely not in
the hands of a programmer.

I looked through a lot of different color pallettes, expanding my
requirements each time. It was difficult to find a pallette that had
sufficient warmth and enough shades of brown, green, and stone-like
colors.

I eventually settled on Resurrect64, a pallette by Kerrie Lake. It
had about 90% of the coverage that I need - including several
varieties of green - but didn't have the coverage of browns that I
needed; nor enough for a darker stone. That's when I gave up on being
a pixel-pallette-purist: if I needed colors, I was just going to add
them myself. The colors ended up working pretty well and, when I
applied the pallette to most of my existing assets, it looked MUCH
better than Fantasy24 did.