Switch from Surt trees to ZRPG trees
I was dissatisfied with the resolution of our old trees (which were by
Surt), so I wanted to replace them.

I have a very specific idea in mind for what I want, but I lack the
artistic talent to bring it to life - and I'm afraid that what I need
might be asking too much for any artist. But I'm going to keep
searching for someone to commission!

However, in the meantime, I needed an intermediate tree to test out
some new possible mechanics I came up with. The ZRPG set by Zabin (but
mostly Jetrel?) had a lot of good trees of appropriate size but not
entirely the correct aesthetic. The pine trees were shorter than I
wanted but were tall enough AND were in the right zone aesthetically,
so I went with those.

This necessitated an update to the "Tree" scene, since this scene is
now only a Pine Tree - ergo, PineTree. Gone are the multiple hitboxes
and the elm-oak-pine dichotomy, and everything is generally simpler.