Update the UnitPawn itself and how it moves
The old UnitPawn movement was a bit wacky. I didn't entirely know what
I was doing, there was some sort of weird slow-down-speed-up thing
happening, and the pawns were constantly overshooting their targets.

I think the design ethos that made for complicated movement code was
that the last thing I had to make move was a servo-motor driven robot,
and I was using some sort of half-remembered feedback systems theory.

I've replaced this code with something that feels like a lateral move
in terms of complexitiy but it works better, is clearer, and has a
whole bunch of comments. We no longer use the move and slide (we'll
have to implement sliding manually) and now do manual checks to see
where the floor is. We also predict how far we'll move and cap out
that movement if it means we'll shoot past our goal.

As a result, the UnitPawns no longer jitter around their goal point
and march at a consistent speed to where they're supposed to go.
Splendid!

While I was updating how movement is handled, I also updated the
UnitPawns themselves - I adjusted the hitbox so it better aligns with
the Clown sprites we're using (and will probably align with our new
sprites, whenever those happen).