Add a Python script for easy dithering
Earlier on in this phase of the project, I was trying to generate
textures that looked like they were dithered but I couldn't find a
good way to do it. I eventually just gave up and moved on.

But for some reason, it was just nibbling at me - and after some
Googling (I remembered the term "convolution") I found a Python
library that ACTUALLY made dithering easy!

I couldn't pass up this opportunity, so I created a really basic and
somewhat obtuse Python script that takes in an image and pumps it back
out in dithered form. It works pretty well and I'm happy with it.

We're gonna keep this tool script in a separate directory, since
there's at least one more tool script I can think of on the horizon.