Add labels to slopetest to observe inner workings
I broke everything in the last commit, but we couldn't observe
anything that was happening. We could only sort of guess and use print
statements. But the print statements came too quickly for us to
discern any appreciable data.

To correct this, I've added a six-label GUI to the SlopeTest. These
are the state, test index, target position, current position, distance
to target, and time remaining. This should allow us to diagnose why
the Pawn isn't moving anywhere...