Organize the Blender-pawn using RAW SCRIPTING POWER
So, last time we created the pawn using a script, but it was
disorganized, left in pieces.

I'm kind of putting off the whole 'animation' thing, and I thought to
myself - why not create a script that can parent everything FOR you?
That was easy enough to accomplish. I was originally just going to
append this onto the end of the generation script, but that script
was already complex enough. Besides, we may use that for other
purposes, including generating gibs.

Now that the body parts are organized into a hierarchy, we just need
to animate them. I think I can write a script that controls the
positioning of each body part, then mix-and-match the functions to
create different overall poses. Then I keyframe those poses and...
that's an animation? I hope so.