Move pawn constants to a neutral file
As a programmer, it was bugging me that the bulk of our constants were
being declared in a single file that was mixed with program code. As
we expand the Blender-Python scripts, we would need to copy-and-paste
these constants over and over again. If even ONE constant differed
from it's fellows it'd throw the whole thing out of whack.

So, I've moved the constants into their own script: pawn_constants.py.
This has all the constant values we were using in generate_pawn.py,
and has some derived constants that we will (probably?) use in other
scripts. It also has the "name" constants so we have a standard way of
referring to the different body components.