Use cylinders and ico-spheres for pawn components
Previously, we were using rectangular prisms for the arms, legs, and
hands of the pawn. This looked pretty good, but presented us with a
pretty big problem.

To prevent the model from looking too uniform, we had rotated the arms
and legs at 45 degrees. However, because the legs were rotated 45
degrees on Z, attempting to rotate the legs or arms on other axes
(axises?) caused the objects to rotate in weird, unpredictable
directions. Technically speaking, since the objects were already at 45
degrees, this behavior was correct BUT it was going to be a BIG
impediment to properly animating this thing.

I could have replaced the arms and legs with rectangular prisms that
were differently constructed, but I honest wasn't craaaaazy about how
they looked originally. So, I decided to hew closer to the reference
model and use cylinders for the arms and legs. I'm very happy with how
that looks.

Since I was updating how things looked, I also changed the hands from
cubes to ico-spheres, which is closer to both the reference model and
my original vision. The ico-spheres do seem to act a bit weird
concerning their diameter but other than that they're pretty solid.