Prepare Pawn for animation using Python scripting
Once we've generated a Pawn, we have all the individual components
that we need to animate. We can use those components for other things
too, like gibs. However, all of the components would be separate,
disparate. If you moved the body would need to move all the other
components in an equal movement.

To fix this, we used the organize_pawn.py to parent all the components
as would be appropriate. But there was still more we needed to do.
Trying to animate the components in this state would be difficult
since none of the pivot points are where you would expect them to be.

So, organize_pawn.py has been changed into prepare_pawn.py. It
maintains the organization functionality, but also now modifies the
origin of (most of) the components so that they'll be easier to
animate.