Create a pose script
We created a pawn, then we prepped it for animation - now we need a
script to explicitly lay out the different poses so that we can get
consistent animations.

To fill this gap, I've added 'pose_pawn.py'. It's pretty standard,
just has a bunch of functions for posing. One thing to note is that I
made a deliberate choice to separate the function by the component
affected - separate for left and right, separate for arms and legs.
By swapping these function calls, we can easily swap out how are legs
and feet are oriented. We can then capture these as key frames and
move to the next pose/keyframe-set.

Also, one of the poses for the feet was nearly clipping the arms in
their "outstretched" state. To compensate, I shortened the feet just a
hair - you can barely notice; it almost looks the same!