Update scripts to run in Blender 2.9
I was using the Blender install you get from Ubuntu, which was Blender
2.79. After some initial confusion in some YouTube tutorials, I
realized that I was using an outdated version of Blender. So I updated
to the latest version, but this created some library
incompatabilities.

To be more specific, the concept of "scenes" had been removed, and
more-or-less replaced by Collections. In addition, the argument path
stopped working and had to be updated.