Add new Pawn sprites
I've finally replaced the stupid clown-pawn-box things with something
that even starts to approach my original vision: featureless isometric
sprites.

Okay so I didn't mean for them to be featureless, that was just
laziness on my part. I haven't really enjoyed working in Blender and
doing the UV skinning would make me scream. So, for now, they're
featureless and animated rather simply.

But they still look pretty good! I was able to generate the model
using the scripts that I wrote, and animate it by mix-and-matching the
poses in the poser script. Thanks to a Youtube tutorial by Surfaced
Studio, I was able to do the most basic level of animation possible to
produce a working sprite. I was then able to set up my scene and
export isometric sprites using a YouTube tutorial by gamevisualz. That
tutorial ended up being a little bit more than I needed but was
fundamental to establishing the scene.

For now, I've colored the sprites to the brightest blue in the
Resurrect64 pallette. I think I'm going to have to do some skinning
eventually, but for now that should suffice.

Since we produced near end-product sprites, I was able to modify the
UnitPawnVomitClown (which I should get around to changing name-wise)
in terms of collision size and code. I have other priorities right
now, but it does mean I can get back to work on movement code since I
have a near-end-product approximation.

To help judge how big the UnitPawn should be, I exported the model
as pawn_rev01 (rev01 since this is the first revision of the model).
This is really only meant to be used for reference since the faces
are kinds of messed up and there are no animations.

Now that the basic sprite movement is done, there's still some more
stuff to do. I need to add a death animation, as well as a dead
animation, since I forgot both of those. I also want to see if I can
somehow change the export settings and miscellaneous scene setup that
went into creating the isometric sprite. That way whoever wants to can
have a ready-made scene that's about 95% ready for export!