Add a dead pose for Revision 01 Pawns
In my haste to get the pawns ready, I forgot to create a death
animation and death pose.

I'm still too lazy to make a death animation - I just took a page out
of old X-COM and made them spin around and fall over. This makes
things easier since I only needed to make one angle for the dead body
pose.

One thing I did notice - our cameras are currently setup with a
purposefully narrow view of the pawn. Initially I thought this was
efficient but as I observed, it left very little room for doing
anything additional with the Pawn. I was able to move the body up so
that it fit in the camera, but this could become a problem if we try
anything more advanced.