Set up the Blender scene via scripting
I've configured the Blender Pawn scene in a very particular way,
using certain lights and certain cameras with certain configurations.
If these scripts are going to be truly portable for end users, then we
need a script that can do the setup for us. That way, any user can
pick up this project and start making animations.

So, I've added a script to do just that. The script is intended to run
after the pawn has been created and positioned - it relies on the
'body' component to center the camera rig. The camera rig itself has
all the lights and cameras you'll need.

The script has several constants that are purely in-script, rather
than in the Pawn Constants - since they weren't related to the Pawn, I
decided they didn't need to be included.

Although we've only really used one camera so far, I decided to
include a head-on camera and a top-down camera. I have ideas about
what we could use these for (pawns sleeping, pawns being carried away
by animals, etc).