Reorganize our scripts
I'm not done making scripts yet, so I figured I should take the time
to organize them. I had originally organized the scripts by
programming language - OpenSCAD and Python.

However, as time went on, it became apparent that this wasn't going to
work. OpenSCAD was fine, but our Python folder had a single utility
script and around six Blender scripts. The Blender scripts couldn't be
run independent of Blender, so including them in the Python directory
was irritating me.

The Python directory has been eliminated, replaced by the Blender
directory. Both OpenSCAD and Blender have been moved under a scripts
directory, which we'll expand as needed.

As part of the clean-up, I deleted two scripts: hither_dither_easy.py
and pyramidion.py. The pyramidion script was meant purely for testing
mesh generation, and didn't really have a use outside of that. The
hither-dither script was difficult to use and wasn't producing the
kinds of dithering that I wanted. I've identified other solutions that
I'm going to be exploring in the near term - the ImageMagick program
looks pretty interesting.

Finally, I added a README.md file to the Blender scripts. This file
explains the purpose of the different scripts, how to run them, what
order to run them, and other caveats. I tried to write the guide with
the perspective of someone who's never done Blender scripting (and
quite possibly next to no programming at all).