Add in a Blender animation script
Previously, I was animating the pawn by editing pose_pawn.py, running
the script, manually keyframing each component, and then repeating.
I'd say this was good but honestly it was so tedious that I only ended
up doing a three-frame walk animation.

There had to be a better way - OF COURSE THERE WAS! I JUST NEED MOAR
SCRIPTS!

So now we have an animation script - pose functions are packaged
together in Pose objects, which can then be packed together into
Animation objects. The script has a utility function for turning the
Animation object into a series of keyframes, complete with animation
repetition and camera rotation.

The sum effect: you get your animation from all 8 angles, and it only
takes the time investment of however long it took you to type all that
crap. Neat!

I've replaced the original spritesheet that I made with a new
spritesheet. I'm happier with the animation, but not too crazy about
it. More work to be done!