Adjust the pawn and it's walk cycle
I was working with the pawns, but I had a couple of issues.

Firstly, the feet were just TOO big. The proportions of the feet were
originally measured from the wooden doll I used as reference. However,
the feet on the doll seem like they were very purposefully designed to
be large and clown-like so that the doll could stand up on it's own.
This turned out to not translate as well to our sprite.

When I was trying to line up the sprite's feet with the ground, the
extra-long clown feet kept clipping into the ground. This forced me to
push the sprite upward, giving it the general appearance of floating,
much to my chagrin.

So, I shortened the pawn's dumb clown feet to a more sensible level,
and then remade the spritesheet. The pawn still looks like it's
floating sometimes but it's not as bad and I don't know how much I
care. I could probably fix via a manual edit that shifts the sprites
until the feet are level, but again: righteous apathy guides my path.

I also revised the walk animation down to 4 frames. The 10 frame
animation never looked that good, so I went back to the original 3
frame animation (the extra frame is another 'idle'/'middle' frame).
This time, the legs dont swing from fullest position, but from the
quarter positions. This looks good enough for me.

Also - since I was feeling lazy, I used the "Simplepacker" sprite
sheet maker, by Maciej Lechowski. However, this tool packs the sprites
in vertical order. So our ordering has now changed: the columns are
one direction of the animation, and the rows each represent the same
frame from different angles.