Goodbye, Artzie Music
I don't think a git repo is where you're supposed to eulogize a
YouTube channel, but I also heard your not supposed to use it to store
binary files - and I've very clearly ignored that. What's one more?

Artzie Music was a Youtube channel that superimposed FutureFunk and
Vaporwave tracks over anime GIFs. It sounds dumb and simple; I guess
on a certain level it was. But that was the beauty of it - Artzie
packaged together music and visuals into a cohesive aesthetic. The
visuals informed the music and the music informed the visuals. Again,
not revolutionary by any means but when I found Artzie Music it was
revolutionary to me.

It's been said that music has a special place on YouTube because it
almost resembles a records store. Songs aren't categorized by artist
or genre, but are instead presented almost entirely on the basis of
their thumbnails. In that environment, Artzie was like a candy store.
Each thumbnail and subsequent GIF was hand-picked - the best designs,
the best colors, the best animation.

That's where I found Vaporwave and FutureFunk. I'm not sure what
about it struck a cord. It just felt like, for the first time, music
actually resonated with me. I'd liked music before, but the music on
Artzie felt like it drew ideas out of me, out of thin air. I'm not
kidding about that - when I found Artzie is when my programming really
kicked into overdrive. Webservers, rendering projects, robots, machine
learning. It was Artzie driving it all. I was even listening to Artzie
Music while working on this project!

And there was this whole community of people who had the same vibe. I
always felt like my music was too weird for most people, and I make a
general point of not mentioning it. But in Artzie there was a great
mass of people who actually liked the same nonsense I did. It really
was the "#1 Destination for Unique Music".

But, there was a hard truth to Artzie Music. A lot of Vaporwave and
FutureFunk really pushed the limit on what constituted sampling and
what was literally just a slowed-down or sped-up version of a song.
And in nearly all cases the songs obscured the songs and artists that
were being sampled from - which seems like a genuine ethical problem.
Frequently, the sampling credits were the top comments on each video,
as everyone came together to try and decipher where and how a song
came to be. And those anime GIFs were presented more or less unedited.
It was a copyright minefield, and a mass DMCA based on those GIFs was
what did Artzie in.

I know why copyright is important. The sampling issues I outlined
above are a really good example of that! I just - I wish my favorite
weird little virtual record store had been spared. That we could have
just had that.

But we had it for a while. And a lot of good artists I first saw on
Artzie are still out there, making music. And Artzie broadened my
horizons. Turns out, a lot of what I liked in Vaporwave and FutureFunk
was in the samples. It was in Disco, R&B, Jazz Fusion, City Pop. And
there are still other channels that do work similar to Artzie, and
there may come another successor (here's hoping)! But it's not a
guarantee, and I guess that's the scary part.

Some people may proclaim the death of FutureFunk, of Vaporwave. To me,
it seems like each "death" (because they apparently die about every
other year or so) is a transition to another phase, another wave of
artists. And it's true that maybe A E S T H E T I C music might die
out as a subculture - the memes and aesthetic principles going towards
the wayside. But I think that what resonated with me originally in
Vaporwave and FutureFunk will resonate with others going forward. I
think there will always be a market for this media, even if the media
can't trace it's heritage back to the starting days of Vaporwave and
FutureFunk.

And besides, the Dodgers won the World Series. Hard to be sad when
that happens.