Test a separate Kinematic driving node
Okay, so the code that drives the Kinematic bodies is pretty
complicated. Worse still, it was almost entirely integrated into
UnitPawn in a way that was difficult to separate.

And that was the real issue - there's gonna be a whole heap of
different things moving around in the game besides the UnitPawn. So I
needed to try and isolate the code that "drives" the nodes into
something that could be easily attached to a kinematic body with ease.

The KinematicDriver script is my proposed solution. Its a more-or-less
duplicate of the UnitPawn's movement code but with a few key
differences. The biggest difference is with how we calculate the
target position - this is handled by whatever node we're attached to
as part of the coupling process. This makes it quick and easy to
attach to any KinematicBody node.

To demonstrate, I added a test/example where I take a simple Node -
called the Cubit - and plugged in the KinematicDriver script. It works
pretty well!