Oh boy: It's a BIG DUMB REFACTOR
Yeah, I was thinking that it was just about time that I royally screw
up the entire history of my git repo by playing a giant dumb game of
musical chairs.

Believe it or not, there's actually a few reasons I wanted to
rearrange the project files.

First off, it was a matter of distance. I had everything divided into
two categories - assets, or scenes. While this was tidy, it meant that
the assets were stored in a parallel directory structure to their
corresponding scene. So to rapidly switch between the scene and the
asset required a lot more movement and navigation than really made
sense. That's the first adjustment we made - grouping together scenes
with their assets. Now this isn't perfect, because sometimes we will
be reaching across for assets, but it's a pretty good solution in
general.

Secondly, the refactor allows me to better position common assets.
This was something I was thinking about when I was creating the
KinematicDriver script - that's a script that needs to apply to a
whole class of different scenes (what I've dubbed "motion AI" because
it's AI that moves). So having a motion_ai directory with a common
subdirectory allows us to better group our KinematicDriver script with
the the scenes that use it. Our common items are now more generally
available.

Finally, as time has gone on, we have experienced a natural accrual of
unecessary assets and structures. A lot of stuff was left from the
Boids prototype, and a lot of the starting directories had been
structured to work around the Boid directories. For example, the
"formation" directory had been acting as a catch-all since we made it,
when it was originally created to differentiate 3D assets from the 2D
Boid assets. Refactoring allows us to remove that useless stuff.