Make UnitPawn step up (and break everything else)
The present struggle in which I was so utterly engaged in fixing was
trying to figure out how to make the stupid UnitPawns step up. They've
been able to slide up slopes, and some low lying platforms - but the
game is going to feature ledges pretty frequently (i.e. for building
foundations and the like).

I was struggling because I needed a configurable solution - not all
units will be able to step up onto foundations or ledges, and some
will be able to step up pretty aggressively. I wanted to try a
solution I saw online that used a RayCast and Capsule solution, but
ended taking a long time to set up and didn't even work.

The solution I came up with is a lot easier to understand - we use a
box shape (once again) and, using a new value in the KinematicDriver,
float a fixed distance above the ground. By configuring our hitbox
correctly, we can easily glide over ledges of a certain size. Once
we're over a ledge, we snap up to the appropriate float height. Voila!
An easy step-up method.

Given that the CollisionShape no longer matches up 100% with the
Sprite, this may have an impact on how we detect attacks (whenever we
get around to that nonsense).

Also, since there's a lot of interplay between the float height, the
collision shape, and the UnitPawn's FLOOR_DISTANCE_ADD variable, it
was getting difficult to configure UnitPawn. So, I added a new test -
the PawnStandingTest. This exciting new test features the Pawn just
standing there, doing nothing. Amazing! .....it's good for
calibrating, okay?

The only problem is that now the UnitPawn climbs all slopes TOO well!
It's like a spider or a monkey! GAAAAH! That's okay, we just need to
do some more normal-angle manipulation nonsense.