Add 3D Labels for debuggery (for now!)
I was having trouble distinguishing the steps and ramps in the
SlopeStepTest. I wasn't having too much of an issue telling the slopes
apart, but I was really struggling with the steps. I want to do some
more fine tuning of the Pawn's step-up/collision model but since they
were essentially unlabeled I couldn't tell where the Pawn was failing.

Since they could change on the fly, I needed something that could be
easily edited. So I've created a class that allows us to move and
manipulate labels attached to 3D quads.

Putting these together was a proper pain, and I'm worried that using
too many might lag the game (since they depend on the Viewport node).
It seems like if you copy them, they don't update right away, and
daring to load the Quad before the Viewport crashes everything, and I
can't really size them or center the text as well as I would want to.

So I basically made a giant pain, but they do have legitimate uses,
especially in debug. And, as we go on, this class will serve as the
template for other 3D GUI elements (like health bars over enemies and
friendly units).

Also, as part of the 3D GUI labels, I added a open source font.