Increase the step-up height for UnitPawns
I wanted the UnitPawns to be able to step up onto ledges that are, at
most, .5 units tall. That's what I'm going to use as the standard for
the foundations of most of the games buildings - armories, houses,
halls, and all the like.

With some adjustment, I got just what I wanted. One thing I noticed
was that the "fake/effective" collision box was actually shifted off
of the origin, which was probably messing with the KinematicDriver's
float-code. I recentered things by placing the full collision model at
origin, adjusting the real collision appropriately, and bringing the
sprite down.