Add "shadow only" mode to most building components
We've had a very particular issue for a while now - when the Pawns
enter buildings, or go behind buildings, we completely lose track of
them. I did actually design around this - that's the reason the
houses & huts are split into lower and upper walls. The idea is we
just "disappear" the upper wall and roof, and then we can see
everybody in the building.

However, that "disappear" functionality had some quirks to it. We
can't just toggle visibility, because the walls still need to cast
shadow. So we need to change the shadow setting for each component
mesh.

Most of the building components - both the simple and composite sets -
now include a "Shadow Only Mode" boolean configurable that allows us
to quickly and easily switch our components from normal mode to shadow
only mode. For composite components, these changes are propagated down
to the subcomponents.

I made the decision to not update the foundation and stairs, since
those will most likely never be used in such a way that they would
obfuscate as walls or roofs would.