Add shadow wall-dropping to the House and the Hut
We needed a way to peek inside our buildings. Since this game uses
isometric projection into 3D, there's cases where things we want to
see (i.e. NPCs, items, player characters, etc.) will enter or go
behind buildings. Ergo, we needed a way to see through the walls.

The first step was ensuring the components could be switched to a
shadow-only mode. That's done, so now we need to implement it at the
building level.

Both the Hut and the House scenes now have interior and exterior area
detectors. They both have logic for detecting what elements to
drop/shadow/collapse.

To help implement this logic, I had to update the WallSet so that we
could selectively hide the different components at will. The boolean
configurable has been replaced by a enumeration of the different wall
states.

Unfortunately, I can't think of a better way to setup the scenes other
than duplicating the script for each scene. The problem is that the
CollisionShapes need to be custom fit for each building (since each
building can have a different footprint). They then need to be
manually tied to the Wallsets and roof, and have the orientation fed
to them. If we end up pretty far along and we have a buncha similar
buildings with the same footprint, I bet I could do something.