Path the Pawns via Navigation!
Previously, we were just feeding the Pawns singular points, which they
then mindlessly marched to. That was regardless of what was in the
way - walls, rocks, ledges, other crap.

Obviously, pathing is a major part of any strategy game; and this
prototype is no exception. We were gonna need a way to feed the Pawns
coherent paths instead of just singular points.

This required a lot of work - not necessarily a rework, but a lot of
small changes.

The first change was in KinematicDriver. The position of a node
navigating a NavMesh needs to be determined by the node and the
NavMesh. In other words, it is beyond the KinematicDriver's ability
to calculate it's current position. To get around this, the
KinematicDriver has a new signal: request_position(). This is a
request for whoever's listening (i.e. the KinematicBody being driven)
to update the Driver's position value. While asking for an update on
a public variable via a signal isn't exactly kosher, I felt strapped
for options. I suppose we could pass Navigation down to the Driver,
but the driver is meant to go from point A to point B. Code should
adapt and specialize around the Driver, not vice versa. Other than
that, the KinematicDriver is largely the same.

The second adaptation was in Pawn (so long, UnitPawn!). The Pawn now
accepts a NodePath to the Navigation node. This Navigation node is
used for checking position, which is then passed to the driver via the
request_position signal. It was important to me that the Pawn maintain
it's old floor-point functionality, so the Pawn switches its
functionality based on whether we have a navigation node provided.

The "destination" in Pawn has been superceded by the "current_path",
which is an array of Vector3 positions for the Pawn to follow. These
are passed piecemeal down to the KinematicDriver as appropriate. The
paths themselves are given wholesale to the Pawn, and are assumed to
be "fine" (i.e. we won't have to edit them). In the future, there will
likely be some sort of HiveMind, PawnManager, or TownManager that will
take care of feeding the Pawn's paths or location data of some kind.

The "ParadeGround" has been deleted. This was always intended as a
demonstration of how the Pawns move and interact with their
environment. It has been effectively replaced by our existing tests
and our latest test - the PawnPathingTest. The PawnPathingTest has
our navigation mesh and sample code for ordering a single Pawn around.

I noticed early on that the NavigationMesh was struggling a lot with
putting down mesh inside the test buildings. It seemed to just LOVE
putting it on the roof. To get around that, I did a rework of our
collision layers.

The first three layers are now all terrain layers. First is Pathable,
which is meant for ground and the like. Next is Nonpathable, which
identifies terrain that needs to be pathed around but shouldn't be
pathed over - like walls. Although I don't have a lot of control of
pathing OVER things, I thought it was important to differentiate. The
last terrain layer is the "Path Ignored" layer, which is meant for
roofs and overhangs that never get accessed (and shouldn't be,
really). This allows us to actually create mesh inside the buildings.
Yay!

The rest of the layers are dedicated to enumerating specific item and
thing types - I'm not sure what we'll need exactly but I'm pretty
confident we'll layers for different groups.

The last four layers are dedicated to "Teams", numbered 1 through 4.
This will be used for friend-versus-foe identification. Eventually.

There were two additional tweaks after observing the Pawn moving.
First, the internal Driver's goal tolerance had to be increased to .2,
since there seemed to be some funkiness happening whenever the Pawn
was stepping up or down. Some sort of vertical mismatch I guess. Oh
well, we don't need it to be too precise. The other problem was that
the Pawn kept catching on edges - so I increased the agent radius
until the clipping stopped.

Now comes the real hard part - doing this all on the fly...